refactor: Add open state to ClassesTable component
This commit adds the `open` state to the `ClassesTable` component in the `ClassesTable.tsx` file. The `open` state is used to control the visibility of a dialog for adding a new class. The state is initialized as `false` using the `useState` hook from React. When the user clicks on the "Ajouter une classe" button, the state is set to `true` using the `setOpen` function. This change enhances the functionality of the `ClassesTable` component by allowing users to add new classes through a dialog.