refactor: Enable CREATE functionality in QUERY_KEYS
This commit enables the CREATE functionality in the QUERY_KEYS object in the query-keys.ts file. The CREATE property is uncommented, allowing users to create a new class by calling the "CREATE_CLASSE" action with the specified options. This change enhances the functionality of the application by providing the ability to create classes through the QUERY_KEYS object.