refactor: Update CreateClasseOptions interface to make description and imageUrl optional
This commit updates the CreateClasseOptions interface in the classe-types.ts file to make the description and imageUrl properties optional. This change allows for more flexibility when creating classes, as not all classes may have a description or an image. It improves the codebase by aligning the interface with the actual usage of these properties in the createClass method and ensures that the options object accurately represents the data needed for creating a class.