refactor: Update CreateStudentOptions interface to make imageUrl optional
This commit updates the CreateStudentOptions interface in the student-types.ts file to make the imageUrl property optional. This change allows for more flexibility when creating students, as not all students may have an image. It improves the codebase by aligning the interface with the actual usage of this property in the createStudent function and ensures that the options object accurately represents the data needed for creating a student.