refactor: Update GradeTypeSelectGroup component to use "Competence-Based Evaluation" as the value for descriptive grades
This commit updates the GradeTypeSelectGroup component in the GradeTypeSelectGroup.tsx file to use "Competence-Based Evaluation" as the value for descriptive grades. Previously, the value was set to "Expertise/Proficiency/To be developed/To be acquired", but it has been changed to provide a more accurate and descriptive label for the grading option. This change improves the clarity and usability of the component by using a more intuitive value for descriptive grades.