refactor: Update StudentsEvaluationTableView to fix table head width issue
This commit updates the StudentsEvaluationTableView component in the StudentsEvaluationTableView.tsx file to fix a table head width issue. Previously, the width of the "Student" table head was set to 200px, but it caused alignment problems with the other table heads. This change removes the width constraint, allowing the table heads to align properly. It improves the visual consistency and usability of the table in the component.