refactor: Update user query in saveUserMutation to use correct field name
This commit updates the user query in the saveUserMutation function of the users.ts file. It replaces the incorrect field name "_id" with the correct field name "userId". This change ensures that the user is correctly retrieved from the database based on the provided userId. It follows the established convention of using the "refactor" prefix for code changes that improve or optimize existing functionality.