feat: Add react-wrap-balancer npm dependency
Add the "react-wrap-balancer" npm dependency to the project. This package provides a balancer component for React applications, which can be used to balance the load between multiple instances of a component. This change introduces a new feature to the project and follows the established convention of using the "feat" prefix for code changes that add new functionality.