feat: Update userSchema with default schoolSubject and country
This commit updates the userSchema in the user-domain/entities/user-schema.ts file. The default value for the schoolSubject field is set to "Arts", and the country field is added with an enum of countries including USA, France, Spain, Germany, UK, and Australia. This change enhances the user schema by providing a default value for schoolSubject and adding support for country selection.